Move _make_cache_key to avoid circular import
